refactor: extract JiraConnectionConfiguration.swift from JiraExportConfig.swift
Splits the `JiraConnectionConfiguration` connection struct out of
JiraExportConfig.swift so that file holds only the export configuration.

Byte-identical move of lines 3-11. Both files import Foundation.

JiraExportConfig.swift: 92 → 82 lines. Tests: 667.
