+++ Sources/BugNarrator/Models/JiraConnectionConfiguration.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+struct JiraConnectionConfiguration: Equatable {
+    let baseURL: URL
+    let email: String
+    let apiToken: String
+
+    var isComplete: Bool {
+        !email.isEmpty && !apiToken.isEmpty
+    }
+}
