+++ Sources/BugNarrator/Models/JiraExportConfig.swift
@@ -1,15 +1,5 @@
 import Foundation
 
-struct JiraConnectionConfiguration: Equatable {
-    let baseURL: URL
-    let email: String
-    let apiToken: String
-
-    var isComplete: Bool {
-        !email.isEmpty && !apiToken.isEmpty
-    }
-}
-
 struct JiraExportConfiguration: Equatable {
     let baseURL: URL
     let email: String
